05.22 11:00 예약일정 완성한 본

+++ src/main/java/com/project/yomozomo/service/ProductListService.java
@@ -52,12 +52,12 @@ public class ProductListService {
     }
 
     @Transactional(readOnly = true)
-    public List<Wishlist> getWishlist(int userId) {
+    public List<Wishlist> getWishlist(Long userId) {
         return wishlistRepo.findByUserId(userId);
     }
 
     @Transactional(readOnly = true)
-    public List<ViewedProduct> getRecentlyViewed(int userId) {
+    public List<ViewedProduct> getRecentlyViewed(Long userId) {
         return viewedRepo.findTop5ByUserIdOrderByViewedAtDesc(userId);
     }
